extract definitions from platform_agent into new module platform_agent_enums, mainly to facilitate integration with mission executive while avoiding issues with circular imports
